Avance del proyecto un 45%

+++ backend/src/main/java/backend/repository/CursoRepository.java
@@ -3,6 +3,6 @@ package backend.repository;
 import backend.model.Curso;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CursoRepository extends JpaRepository<Curso, Long> {
+public interface CursoRepository extends JpaRepository<Curso, Integer> {
 
 }
