+++ backend/src/main/java/backend/repository/EstudianteRepository.java
@@ -3,6 +3,5 @@ package backend.repository;
 import backend.model.Estudiante;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EstudianteRepository extends JpaRepository<Estudiante, Long> {
-
+public interface EstudianteRepository extends JpaRepository<Estudiante, Integer> {
 }
