+++ backend/src/main/java/backend/repository/MatriculaRepository.java
@@ -3,6 +3,5 @@ package backend.repository;
 import backend.model.Matricula;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MatriculaRepository extends JpaRepository<Matricula, Long> {
-
+public interface MatriculaRepository extends JpaRepository<Matricula, Integer> {
 }
